centeralize nuxt startup requirements
